% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_dir}
\alias{cache_dir}
\title{Get Retrosheet Cache Directory}
\usage{
cache_dir(create = TRUE)
}
\arguments{
\item{create}{Logical. If TRUE (default), creates the directory if it
doesn't exist.}
}
\value{
Character string with the cache directory path
}
\description{
Returns the path to the retrosheetshow cache directory where downloaded
files are stored.
}
\details{
The cache directory is determined by \code{tools::R_user_dir("retrosheetshow", "cache")}.
On most systems this will be:
\itemize{
\item macOS: \verb{~/Library/Caches/org.R-project.R/R/retrosheetshow}
\item Linux: \verb{~/.cache/R/retrosheetshow}
\item Windows: \verb{C:/Users/<user>/AppData/Local/R/cache/R/retrosheetshow}
}
}
\examples{
\dontrun{
# Get cache directory
cache_dir()

# List cached files
list.files(cache_dir())
}

}
