% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_status}
\alias{cache_status}
\title{Show Cache Status}
\usage{
cache_status()
}
\value{
A tibble with information about cached files
}
\description{
Displays information about the retrosheetshow cache, including size and
contents.
}
\examples{
\dontrun{
# View cache status
cache_status()
}

}
