% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_event_url}
\alias{construct_event_url}
\title{Construct Retrosheet Event File URLs}
\usage{
construct_event_url(year, type = "regular")
}
\arguments{
\item{year}{Year of the data}

\item{type}{Type of events ("regular", "allstar", "post")}
}
\description{
Construct Retrosheet Event File URLs
}
\keyword{internal}
