% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for a given event type}
\usage{
get_available_years(type = "regular")
}
\description{
Get available years for a given event type
}
\keyword{internal}
