% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{parse_event_file}
\alias{parse_event_file}
\title{Parse Retrosheet Event File Lines}
\usage{
parse_event_file(lines, year)
}
\arguments{
\item{lines}{Character vector of lines from event file}

\item{year}{Year of the data}
}
\description{
Parse Retrosheet Event File Lines
}
\keyword{internal}
