% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_events.R
\name{parse_record_by_type}
\alias{parse_record_by_type}
\title{Parse individual record based on type}
\usage{
parse_record_by_type(type, fields)
}
\description{
Parse individual record based on type
}
\keyword{internal}
