% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrosheetshow-package.R
\docType{package}
\name{retrosheetshow-package}
\alias{retrosheetshow}
\alias{retrosheetshow-package}
\title{retrosheetshow: Access and Parse Retrosheet Baseball Data}
\description{
Provides a convenient and tidy interface for accessing Retrosheet baseball
data, including play-by-play event files, game logs, rosters, schedules,
and reference data. Functions follow tidyverse principles for easy
integration into data analysis workflows.
}
\section{Main Functions}{


\strong{Events (Play-by-Play)}
\itemize{
\item \code{\link[=list_events]{list_events()}}, \code{\link[=get_events]{get_events()}}, \code{\link[=get_plays]{get_plays()}}, \code{\link[=get_game_info]{get_game_info()}}
}

\strong{Game Logs (Summary Stats)}
\itemize{
\item \code{\link[=list_gamelogs]{list_gamelogs()}}, \code{\link[=get_gamelogs]{get_gamelogs()}}, \code{\link[=gamelog_fields]{gamelog_fields()}}
}

\strong{Rosters and Schedules}
\itemize{
\item \code{\link[=get_rosters]{get_rosters()}}, \code{\link[=list_schedules]{list_schedules()}}, \code{\link[=get_schedules]{get_schedules()}}
}

\strong{Reference Data}
\itemize{
\item \code{\link[=get_park_ids]{get_park_ids()}}, \code{\link[=get_team_ids]{get_team_ids()}}, \code{\link[=get_player_ids]{get_player_ids()}}
}

\strong{Cache Management}
\itemize{
\item \code{\link[=cache_status]{cache_status()}}, \code{\link[=clear_cache]{clear_cache()}}, \code{\link[=use_cache]{use_cache()}}
}
}

\section{Retrosheet Data Notice}{


\strong{IMPORTANT:} The information used here was obtained free of charge from
and is copyrighted by Retrosheet. Interested parties may contact Retrosheet
at 20 Sunset Rd., Newark, DE 19711.

Website: \url{https://www.retrosheet.org}
}

\section{Getting Started}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{library(retrosheetshow)

# Download play-by-play events
events <- get_events(year = 2024)

# Get game logs (faster, summary stats)
gamelogs <- get_gamelogs(year = 2024)

# Get team rosters
rosters <- get_rosters(year = 2024, team = "NYA")
}\if{html}{\out{</div>}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://tgerke.github.io/retrosheetshow/}
  \item \url{https://github.com/tgerke/retrosheetshow}
  \item Report bugs at \url{https://github.com/tgerke/retrosheetshow/issues}
}

}
\author{
\strong{Maintainer}: Travis Gerke \email{tgerke@mail.harvard.edu}

}
\keyword{internal}
